Refresh process list as soon as it becomes visible
Summary:
`setInterval` calls its argument after waiting some specified number of milliseconds.  We want it to
be called right away, as well as after each subsequent delay.  If we don't do this, there can be old processes
in the list that hang out.  I was testing the node debugger and kept trying to attach to dead processes
because of this.

Changelog: Node and Native debugger process selection lists refresh as soon as the dialog is opened.

Differential Revision: D4436145

fbshipit-source-id: ad6d334b04b02798632ea01d2758b2871b5ab09c