Turn off debugger extensions causing bad behavior
Summary:
Node inspector lets you inject "debugger extensions" which is a feature we don't use.  This feature is on by default, but was failing, and this failure prevented a loader breakpoint pause from being sent to Nuclide when the debugger attached to the process.  The loader bp is an important part of our state machine, and the debugger expects it.  If it's not sent, the UI will look a though the debugger is in run mode, when it is actually in pause mode.

Disabling this config option fixes the UI, whereas before the node debugger was pretty unusable.

Changelog: Bugfixes for attaching the Node debugger to a Node process.

Differential Revision: D4438836

fbshipit-source-id: fda2fec14c932d393e216343d774890a57957858