try to kill most of the processes when killing a process tree
Summary:
When killing a process tree, it might happen that a process in the tree does not exist any more, which throws an ESRCH error and the processes that were not tried to be killed will remain there.
Instead, i'm trying to kill them all and after than throw an error if any.

Reviewed By: AsyncDBConnMarkedDownDBException

Differential Revision: D4438066

fbshipit-source-id: 1555fb3871718528436f294e9692c08ed277c40c