Add run-package-specs atom-script and shell script
Summary:
This adds an `atom-script` script that opens a spec file in the spec window without having to load all of Nuclide. Also added a shortcut script as `scripts/run-package-specs.sh`. To use it, just run `scripts/run-package-specs.sh` with a path to a spec file, then it'll open up the spec window with that test loaded. Example:

```
./scripts/run-package-specs.sh spec/quick-open-provider-cycle-integration-spec.js

./scripts/run-package-specs.sh pkg/nuclide-quick-open/spec/SearchResultManager-spec.js
```

Reviewed By: AsyncDBConnMarkedDownDBException

Differential Revision: D4444976

fbshipit-source-id: fab986eb9b6df5751b4844ef8a14f134572f68ae