Try not to complete callable parameters if they are already there
Summary:
The real fix would be Hack-server side and use of full fidelity AST to figure out which parameters are there to suggest a full or partial completion.

For now, using a heuristic: complete only the function name, omitting the params, if the first non-whitespace character after the completion point is "(".

Reviewed By: nmote

Differential Revision: D4439375

fbshipit-source-id: 047affec840ea5474dbdbeaafde99e99881f59be