Move calls to `editor` instance methods above await
Summary:
We do not know how long `await` will take, and `editor` is mutable. If it takes
long enough for the Promise to resolve, the cursor position may have changed by
the time we ask for it. That is setting us up for a strange and difficult to
reproduce/debug issue.

Reviewed By: peterhal

Differential Revision: D4458516

fbshipit-source-id: f11c570b1b9a1158e657e7801ddf4a89f198e5ad