Fix sticking panel resize on Windows
Summary:
So, for some reasons there's this weird behavior on Windows, when you start resizing a panel, but are never able to stop. Even as you release the mouse button the panel just keeps on resizing and the only way of fixing it is reloading Nuclide (or quitting Atom).
I've seen this before, but then this suddenly stopped happening. Today I was able to reproduce it consistently and it is weird - the window 'mouseup' event is just not firing. I have no idea why.
To remedy this I'm adding this small check to the mouse move so make the pane resize "unstick" as it's impossible to work in this mode.

I have no idea what is making this behavior inconsistent, but it obviously is, otherwise we'd receive a shitload of issues about it, whereas I vaguely remember two. Both of which I couldn't reproduce at that moment.

Reviewed By: matthewwithanm

Differential Revision: D4459680

fbshipit-source-id: b3d7b029dc21a7fd5be83fe46e9fbf650d659a96