Ensure clicking selected callframe navigates to correct file/line position
Summary:
I tried to add this as an optimization but forgot about the case where the user has navigated to a
different file than the selected callframe.  In this case, clicking the selected callframe should navigate
back to the file/line the user expects.

Changelog: Clicking the selected frame in the debugger callstack window navigates to the expected file.

Differential Revision: D4475032

fbshipit-source-id: b22623bc3b08b814b6becdfb1ef3477dfddb1232