Switch dropdown from tasks to task runners, add individual task buttons
Summary:
Change from task lists to a "task runner state" which encompasses whether the task runner is relevant to the current working root at all, and what tasks it has available. Tasks are rendered as individual buttons, the dropdown only has options for individual task runners. Changed rules around visibility as well – tldr is we remember whether the toolbar should be shown for a specific working root and what task runner should be selected in the dropdown.

{F65566604}

{F65566614}

{F65566617}

{F65566619}

{F65566621}

{F65566625}

{F65566627}

{F65566637}

{F65566638}

{F65566639}

Reviewed By: matthewwithanm

Differential Revision: D4474976

fbshipit-source-id: 9107807bff3b9d513d9bb95d48d42942f75284cd