Fix Hack syntax highlighting for function types
Summary: function highlighting in atom/nuclide for hack language broken when function definition does not end with curly bracket {. Added additional end match for function highlight to stop searching when encountering a colon or semi-colon to cover use cases were function is called from interface or when type hint applied to function call. This prevents highlight rule from failing to end causing the rest of the highlighting to fail.

Reviewed By: dabek

Differential Revision: D4485456

fbshipit-source-id: 1459b2cd5b53689a4ef5bd53d8021ba1fa161c5a