Add a service that provides the compilation db file in the client side
Summary:
This will be useful for certain projects to prepare the compilation database before calling the clang rpc.
I didn't implement a rebulid functionality because right now we don't need it.

On the rpc side we now cache the pair source-compilation_db_path instead of just the source. In case the same source
is used for different compilation db.

Reviewed By: hansonw

Differential Revision: D4499576

fbshipit-source-id: dbf819259a682d198c027ab1c20e756fce0dc9e5