Move much of the Flow diagnostics machinery to the server
Summary:
This leaves us in a strange intermediate state. But I tried to do it all at once and became confused. This is a good step towards the goal, which is to move all of the Flow-specific diagnostics logic over to the server side, and then make Flow diagnostics go through the language service.

I will continue cleaning things up in upcoming diffs, so please do not focus too much on the weird details, like the temporary `NewDiagnostics` type.

Reviewed By: zertosh

Differential Revision: D4496389

fbshipit-source-id: 9a2ec3baa8f7a3931304b65d415f9e8ac190be86