Detect if requested Android activity exists, and launch under debugger.
Summary: This change first queries the attached device to see if the requested package + activity exists, so we can give a useful error message if the user mistypes. It also modifies the flags passed to `am start` to start the package with debugging enabled, rather than issuing a separate am command. This gives the desired behavior where the app starts frozen with a 'Waiting for debugger' Android system overlay until the debugger finishes bootstrapping.

Reviewed By: johnislarry

Differential Revision: D4512066

fbshipit-source-id: a5b48b6138efb666565381cfa29e98d898fa6b06