Show task runners for new working roots
Summary: Some new users were complaining the experience feels less "magical" since the task runner toolbar doesn't show up automatically anymore. I imagine this will also annoy other people who don't want to use the toolbar, but since this is a one-time thing on a working-root basis, it should be okay (the visibility preference is stored afterwards). Imo the ultimate solution to this might be to have .nuclide files checked into a directory determining the default behavior for the repo.

Reviewed By: matthewwithanm

Differential Revision: D4522879

fbshipit-source-id: 1341bbb756c35f0390ff03013b0564b1d946876d