End debug session when proxy is killed
Summary:
The debug proxy can go away if it's killed or encounters an error.  Nuclide swallows this, but we
should report it to the user.  Right now this diff will just close the debugger when this happens, and write
to a log.  I'll add a channel to send atom notifications back to the user to actual make this error more
user-friendly

Changelog: Mobile JS debugger will exit when debug proxy is killed.

Differential Revision: D4525798

fbshipit-source-id: 906451894e6955d90841f6de91fe27f648eecdff