Add notification channel to ClientCallback
Summary:
This exists for the PHP debugger but it's written in an ad hoc way.  This adds a channel to the
generic ClientCallback class, which supports sending atom notifications back to the client.  Example usage is
in the iwdp package, which uses this channel to send a warning notification when the debug proxy is
unavailable.

Differential Revision: D4529990

fbshipit-source-id: dad67b304e027f63e0599ec461ef0d0c493d553e