Fix nasty bug in runCommand()
Summary:
Previously, we mutated the seed (which was shared between subscriptions of the returned observable). That meant that the output from each execution would be appended to the output of the previous one(s). o_O

While I was in there, I also made it not concat the strings each time.

Reviewed By: shushz

Differential Revision: D4540009

fbshipit-source-id: 728ecd7023f96c529b8cc45294a7eaedb76cf1bb