Add ability to get a list of JDWP targets from an Android device
Summary:
This change adds methods to the AdbService that will query an attached Android device and determine a list of  Java processes that are valid JDWP targets by combining the output of `adb jdwp` with `adb ps` and cross-referencing the two lists to determine PID, process name and process user.

These will be the valid targets that the Android Java debugger can connect to on the attached adb device.

Reviewed By: johnislarry

Differential Revision: D4518548

fbshipit-source-id: 98520fdea0fa0683f948d4b062603c1c7f0e327e