Diagnostics: Add toggle to show full diagnostic traces
Summary:
nmote In response to #998 cc ColinCampbell

It's common to need more information to debug flow errors. This change reduces the need to open terminal and run ``flow`` and doesn't clutter the interface.

Settings still has a "Show full diagnostic traces" option. It's used as a default value for when the "Diagnostics" view is constructed.

![showinteraction](https://cloud.githubusercontent.com/assets/16157429/22670054/56ac6a0a-ec7c-11e6-85b8-b8184a84507b.gif)
Closes https://github.com/facebook/nuclide/pull/1017

Reviewed By: matthewwithanm

Differential Revision: D4523199

Pulled By: nmote

fbshipit-source-id: 1af21404040a94cbd2b2001f5d57c06093311c86