Add assertion that ServerLanguageService satisfies LanguageService
Summary:
There are so many types and classes similar to LanguageService that I have
trouble keeping them all straight. The purpose of this diff is mostly
documentation. Because ServerLanguageService instances are already used where
values of type LanguageService are expected, Flow already enforces this. But
it's helpful to me as a reader to have this relationship written down near the
definition.

Reviewed By: peterhal

Differential Revision: D4546871

fbshipit-source-id: 298c62d46512ddaf20ff8f29e2fbd616b1f77b1a