Add more logging, disable too-noisy logging
Summary:
There is a bug (#15707104) that results in client unsubscribing from server's observable diagnostics. There was more logging enabled in D4424333, but it is so spammy that it didn't produce a useful report - unless the users notices and reports the bug in few minutes since it occurred, the relevant logs are already recycled by then. In the cases where (I think) relevant time-frame was included, the logs are seem to lack useful info.  To try to get some more info, in this diff I'm:

- setting the default log level to DEBUG
- downgrading the too-noisy loggers to TRACE
- upgrading some TRACE ones that we still care about to DEBUG (which, when using CategoryLogger, for some reason is just logger.log)
- adding ERROR logs in some of the places that currently silently swallow errors

Reviewed By: peterhal

Differential Revision: D4547281

fbshipit-source-id: 1ae3244d997f38f169b92d29f94e3ea124adae4c