Show "(no variables)" in Scopes window for frames with no Locals
Summary: Currently if you switch to a frame that has no locals (either because there aren't any or there are no symbolss), scopes shows a weird collapsable container for "Locals" with no content in it. It almost looks like it failed to render locals or something.  This just adds a bit of text to indicate there really is nothing to show.

Reviewed By: johnislarry

Differential Revision: D4574758

fbshipit-source-id: 98511d8354fba68faecb8760d9f55cc62bc114a7