Fix nuclideUri.uriToNuclideUri returns malformed file paths on Windows
Summary: This routine corrupts file:// URIs on Windows. On Win32, something like file://C:/Users/.../some/file (or even file://C:\\Users\\...\\some\\file !) is a valid file protocol URI to a local file. When this routine hands the string to url.parse(), parse() sees the "C:" and thinks the hostname is "C" and loses the ":" entirely. This results in corrupt paths that don't work. Seeing this problem when talking to a local Java debugger RPC server on Windows.

Reviewed By: johnislarry

Differential Revision: D4577478

fbshipit-source-id: 1d25cf50d13bb11c52f35a6ebf53db80481b7636