Update autocomplete-cacher to allow for skipped characters between autocomplete requests
Summary:
By default, autocomplete-plus has a debouncing period where it will not
send an autocomplete request until there have been no events for 100 ms. This
means that when you type very fast, you may not get a request for every
character. Until this diff, AutocompleteCacher assumed that it would get a
request for every character, and so when this was not the case, it would fail to
cache and would instead re-query the underlying provider. This was not a
correctness problem but it was a case of AutocompleteCacher being overly
conservative.

Reviewed By: hansonw

Differential Revision: D4594425

fbshipit-source-id: bcb33336397c61e1ee7a9c84d0176d0125e09852