Only hide debugger pane when user explicitly asks
Summary:
Building off ebluestein's work with fixing the "Stopped" state ui of the debugger pane.  We now only
hide the pane on behalf of the user if they explicitly request it, which can only be done by:

1. The command palette action
2. Hitting the big red 'x' in the top right
3. Hitting the 'x' in the debugger pane tab

Actions that used to involve hiding the debugger pane that now do not do this include:

1. Hitting the 'stop' button in the debugger controls
2. Launching a script under the debugger that runs to completion

In these cases, we simply leave the pane open and the state is switched to 'stopped' which greys out un-usable
UI, and prompts the user to start another debugger session.

Reviewed By: ebluestein

Differential Revision: D4594481

fbshipit-source-id: 5a65f9d2ab31306efee695c1a77e02df109706dd