Prevent commit/amend/publish message getting lost on failed publish
Summary: If the publish step fails the update message gets lost as the state of diff view gets set to BROWSE mode, the publish flow already has a special mode where on failure it doens't clear out the diff view state. This change makes sure that if publish is chosen the flow is handed over to the publish method after commit/amend so that in case of failure the data isn't lost.

Reviewed By: vjeux

Differential Revision: D4599640

fbshipit-source-id: 5f7b50f73fe14507b872b5caac957e0255f690ae