Fix chrome protocol message id mapping
Summary:
The ids that the client sends and that we send to the device are different because we have to
multiplex.  I missed these last time, so I'm adding them in now.  This is pretty gross and error prone and I'd
like to eventually refactor the multiplexer to automatically handle this mapping.

Reviewed By: ebluestein

Differential Revision: D4600566

fbshipit-source-id: 25f98afbfa27095fbebd0fbd6f23c599fdbbd6b6