Print actual buck command to console
Summary: For the sake of transparency and reproducibility, it seems like a good idea to always print the actual command buck service is executing, so the user can copy it and run manually in case they want to diagnose a problem. Among other flags it includes `--config client.id=nuclide`, I could potentially filter that out if we want it to be less noisy.

Reviewed By: hansonw

Differential Revision: D4601325

fbshipit-source-id: cadb63aaab0e9c6840358c8e4aa9296892a132c9