Debugging apple_test build rules
Summary: Enables the Debug task for apple_test targets. The UX is still a bit wonky when cancelling the task, because it's a single buck process, similarly to buck run, will follow on polishing that. It does indeed enable a flow not available before, though. The .buckversion will need to be bumped before this works.

Reviewed By: hansonw

Differential Revision: D4600527

fbshipit-source-id: 986ac0da6be78998ccbe9954d55675d2f8897326