Preserve QuickSelectionComponent across quick open invocations
Summary: Apparently being able to undo in Quick Open's query input across multiple QO invocations is important. As part of the last refactor of QO, I made it so that `QuickSelectionComponent` was fully unmounted when QO became hidden. This disposed of the query input's `AtomInput`, so there went the undo history. This diff makes it so `QuickSelectionComponent` is never unmounted across invocations (only at Nuclide's deactivate), but event handlers are added/removed whenever it's hidden.

Reviewed By: ebluestein

Differential Revision: D4617156

fbshipit-source-id: 754103e31b96303469d1322ec629621f5eac6ffc