Fix circular JSON structure stringification error
Summary:
Run a failing `buck test <something>` and observe the error printed to the dev console. I have serious doubts about the usefulness of this error in general, as tests are to be expected to fail, but we definitely should not try to serialize circular structures while logging the failure.
This was caused by one of the changes in D4474976
I didn't analyze whether there are any other similar problems, so feel free to reject in favor of a more complete solution.

Reviewed By: stepanhruda

Differential Revision: D4621163

fbshipit-source-id: 57a2a86330a4731b5538aef3e027917f10fe49a7