Respect new shouldDestroyOnFileDelete setting in Atom 1.15 for remote files
Summary:
This modifies `NuclideTextBuffer` to support the new `shouldDestroyOnFileDelete` parameter, if it exists, which prevents files closing on delete (a common complaint).

See https://github.com/atom/text-buffer/pull/178.

The implementation here slightly differs in that we don't set the path to null.. but IMO this is a better behaviour, as this way you can just save the file again to preserve it.

Reviewed By: matthewwithanm

Differential Revision: D4616627

fbshipit-source-id: f6258f0bd11801ccfd226720d302ed9a5ace66d8