Bump log4js version to 1.1.1
Summary:
This fixes a critical bug where log4js would throw if the logging directory was removed from under it. This can happen for idle Nuclide servers that get re-connected to, for example, and would trigger a crash.

There's no API difference as far as I can tell.

The new `date-format` and `streamroller` dependencies are just parts of `log4js` that the author separated out, so it's nothing significant.

Reviewed By: vjeux

Differential Revision: D4641396

fbshipit-source-id: 8790a6156df69e73bf44b245c7e44e665be853d5