Wire up to PHP scripts [2/2]
Summary:
HackAST is shipped as a script in the www directory, which we'll shell out to. It's kind of slow at the moment, but the `php -s` flag, which will run the script on the user's active webserver, should make this nearly instant.

The flag isn't enabled yet, but I hear it will be soon.

Reviewed By: vjeux

Differential Revision: D4652746

fbshipit-source-id: 6722f1114955b1073125870a1c827c420d6fbf92