Create commons-atom/projects helpers to get Directory/File objects for a given path
Summary:
The "find references" UI currently uses the FileSystemService to obtain its file contents.
This makes it difficult to refactor out of Nuclide, since it depends on the remoting infra.

Instead, let's create some helpers to obtain Atom `Directory` and `File` objects, which are local/remote agnostic.
We can then use `File.read` in lieu of `FileSystemService.readFile`.

There's a small performance overhead of creating the file object and hashing its contents, but it should be OK for this use case.

Reviewed By: matthewwithanm

Differential Revision: D4663777

fbshipit-source-id: fdb5766209211d622215f41bbdc4c433f8ab1b51