Fix scrolling while defining Working Set
Summary:
It seems that Chrome is preventing from a focused element to leave the view, which becomes a problem during the Working Set definition - the file tree cannot be scrolled properly.
This diff prevents the Working Set defining checkboxes from receiving focus by preventing the default action in their `onMouseDown`.

Reviewed By: matthewwithanm

Differential Revision: D4270200

fbshipit-source-id: 335ba466be05e07545487dfb82f69b851ac31c0d