Show loading spinner while waiting for debugger async break
Summary:
When you break the debugger (click the pause button), it often takes a little while for the debugger to break in. This is especially noticable when async-break'ing into a Hack webserver - it can take like 30+ seconds to come back. Currently after you click "pause", the UI doesn't react until the debugger sends the paused notification. This often results in me clicking the button multiple times, which then resumes my target as soon as it pauses.

This change disables the button after clicking it once and replaces the "pause" icon with a tiny loading spinner to indicate we are working on it.

Reviewed By: jgebhardt

Differential Revision: D4664772

fbshipit-source-id: a89e3c1fb7c9a854a4858872384d89804907c342