Show notification when unavailable instead of a permanent label
Summary:
I noticed that the message would show up when the toolbar was getting ready after launch, which we want to definitely avoid. Let's instead show a notification if the user explicitly decides to show the toolbar, but there is nothing to show.

Note that half of this diff is a backout of D4706525

Depends on D4714752

Reviewed By: matthewwithanm

Differential Revision: D4714875

fbshipit-source-id: a5ec2bdaf6512b1f773079adbaf8067257b1c790