Remote Atom: write state to ~/local rather than /tmp
Summary:
Writing the remote Atom state to /tmp isn't great because it can easily get wiped if tempfiles are periodically wiped. It also means multiple users on the same machine will lock each other out.

We also don't want to pollute the user's homedir, so as a tradeoff try `~/local` if it exists first.

Reviewed By: zertosh

Differential Revision: D4717249

fbshipit-source-id: 4ca6281753551524366b71593bd06d4be8132939