Always ping the Flow server when it is not in a steady state
Summary:
The limit on the number of pings caused issues with starting the persistent
connection. With this change, we will always be pinging the Flow server when it
is in a non-steady state, so we can't get stuck in the situation where the IDE
connection is waiting for the Flow server to be ready, but we aren't actually
checking whether the Flow server is ready.

Reviewed By: hansonw

Differential Revision: D4724074

fbshipit-source-id: 29fd34d0c42e55593c8f49297b136109af31d7fe