Add DisplayableRecord type to represent a Record with metadata to be retained for display purposes
Summary:
To achieve windowing, we need to know the height of each record in the console. Unfortunately this height can change and is not uniform across all records, so we need to store it somewhere.

This information is tied to the rendering of a record, and so it cannot live in the store because there may be multiple consoles and the same record can have a different height between the two console panes (for example, if it's an expandable record and its state is different between the two panes). So instead it is treated as metadata that is stored with the record in the `ConsoleContainer`. When records are appended or removed, the `toDisplayableRecords()` function helps keep the expansion state and height consistent for the existing records using referential equality on the record object. A dedicated `expansionStateId` needs to be used as `RecordView`s will be ephemeral with windowing, and we cannot use the `record` itself because the expansion state is not shared across different consoles.

Reviewed By: matthewwithanm

Differential Revision: D4654179

fbshipit-source-id: c2c8aed1cd6f024e6c5221b87082fee8418d08da