Put stronger typing into quick-open SearchResultsManager
Summary:
This is a small diff in preparation for future work. It separates directory providers from global providers more thoroughly. I also renamed variables where they were confusing.

For keeping track of global providers, I called the member "globalEligibleProviders". As of the current diff, every global provider is automatically eligible. In a future diff, I'll add the ability for global providers to say whether they're eligible or not.

TO NOTE: previously, directory providers only worked after calling "updateDirectories" (for when the set of top level directories has changed), but global providers worked even without that.

After this change, global providers are now similar to directory providers in that they only work after a call to updateDirectories.

Is this a race condition that was manifest for directory providers before and is now manifest for global providers as well? Or was it never a race? I don't know. I haven't observed any problems...

Reviewed By: hansonw

Differential Revision: D4720529

fbshipit-source-id: 4d03b7817e9fb5ad3195f8f290e96afcf533c3e8