Add concept of "eligibility" for global providers
Summary: I'm adding a concept of "eligibility" to global providers. They will base their eligibility on the list of top-level directories (aka projects) that are open. The end goal is so that the Symbols global provider can be shown if there are any languages (Hack or LSP) that support project-wide symbol search.

Reviewed By: hansonw

Differential Revision: D4720715

fbshipit-source-id: 346c7ec77c7d36e9f8f88e6fd5b7c587576d2683