Public signature of HackLanguage no longer depends on HackService
Summary:
We wish to move Hack away from using HackService, and towards using LSP -- i.e. on the server side, we will no longer instantiate a HackServiceImpl, and instead we'll instantiate a PerConnectionLanguageService.

As a transition step towards that, I want the public signature of HackLanguage to expose itself only as the common supertype "LanguageService", and then I can use a gatekeeper under the hood to decide whether to instantiate HackServiceImpl or PerConnectionLanguageService.

Reviewed By: nmote

Differential Revision: D4721433

fbshipit-source-id: 5420264da73a5796c20c56257d2efd8d6ce0e992