AutocompleteCacher: ignore cache if updateResults returns null
Summary:
Allow users of the cache to return null from `updateResults`. This acts as a way to 'cancel' the cache dynamically without having to return null.

As we discussed, ideally the user's `getSuggestions` should return an object with an additional `isComplete` parameter to be a bit more obvious, but this solves the issue for now.

Reviewed By: nmote

Differential Revision: D4760255

fbshipit-source-id: af1342118dac05ed4ea9c76d1b1952c39cf91258