JS number overflow in debuggers causing incorrect output
Summary: Numbers that exceed max number in JS in console REPL or Watch Expressions for all debuggers are printing incorrect values. The fix is to have the Chrome devtools leave numbers as strings. We'll still format them correctly in the console because that is based on the declared type property of the JSON object.

Reviewed By: vjeux

Differential Revision: D4774053

fbshipit-source-id: ce1fc167436c80bcad529a06a351ec4902a1c53e